Move Editor and CodeEditor to common
