+++ source/win-main/App.vue
@@ -61,7 +61,7 @@ import Sidebar from './Sidebar'
 import Tabs from './Tabs'
 import SplitView from '../common/vue/window/SplitView'
 import GlobalSearch from './GlobalSearch'
-import Editor from './Editor'
+import Editor from '../common/vue/Editor'
 import PopoverExport from './PopoverExport'
 import PopoverStats from './PopoverStats'
 import PopoverPomodoro from './PopoverPomodoro'
